+++ src/Elements/Generate/GeometricElement.g.cs
@@ -54,7 +54,7 @@ namespace Elements
     
         /// <summary>When true, this element will act as the base defintion for element instances, and will not appear in visual output.</summary>
         [Newtonsoft.Json.JsonProperty("IsElementDefinition", Required = Newtonsoft.Json.Required.DisallowNull, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public bool IsElementDefinition { get; } = false;
+        public bool IsElementDefinition { get; set; } = false;
     
     
     }